Gig-Data save in database

+++ GigHub/ViewModels/GigFormViewModel.cs
@@ -7,7 +7,7 @@ namespace GigHub.ViewModels
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int GenreId { get; set; }
+        public byte GenreId { get; set; }
         public IEnumerable<Models.Genre> Genres { get; set; }
     }
 }
